requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.7
pdfplumber==0.9.0
python-docx==0.8.11
spacy==3.6.1
nltk==3.8.1
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
phonenumbers==8.13.19
scikit-learn==1.3.0